Set up kernel paging, not works in user program

+++ kern/signal.c
@@ -42,9 +42,9 @@ void signal_jump(void *trapframe, void (*handler)()) {
     current->signal_context = signal_context;
 
     // return address save in x30 in ARM
-    tf->x[30]   = signal_return;
-    tf->elr_el1 = handler;
-    tf->sp_el0  = signal_context->stk_addr;
+    tf->x[30]   = (long)signal_return;
+    tf->elr_el1 = (long)handler;
+    tf->sp_el0  = (long)signal_context->stk_addr;
 }
 
 void signal_back(void *trapframe) {
